move eval to its own mod
